Simple probes working in 2D/serial. Needs testing in parallel and 3D extension
