+++ pyranda/__init__.py
@@ -2,3 +2,4 @@ from .pyranda import pyrandaSim
 from .pyrandaIBM import pyrandaIBM
 from .pyrandaTimestep import pyrandaTimestep
 from .pyrandaBC import pyrandaBC
+from .pyrandaProbes import pyrandaProbes
